refactor(go.d/snmp_topology): organize root package files
