+++ src/go/plugin/go.d/collector/snmp_topology/collector_test.go
