+++ src/go/plugin/go.d/collector/snmp_topology/topology_cache.go
@@ -127,11 +127,6 @@ type stpPortEntry struct {
 	designatedPort   string
 }
 
-type topologyVLANContext struct {
-	vlanID   string
-	vlanName string
-}
-
 type arpEntry struct {
 	ifIndex  string
 	ifName   string
