+++ src/go/plugin/go.d/collector/snmp_topology/topology_lldp.go
@@ -4,6 +4,33 @@ package snmptopology
 
 import "github.com/netdata/netdata/go/plugins/plugin/go.d/collector/snmp_topology/internal/topologyutil"
 
+var lldpChassisIDSubtypeMap = map[string]string{
+	"1": "chassisComponent",
+	"2": "interfaceAlias",
+	"3": "portComponent",
+	"4": "macAddress",
+	"5": "networkAddress",
+	"6": "interfaceName",
+	"7": "local",
+}
+
+var lldpPortIDSubtypeMap = map[string]string{
+	"1": "interfaceAlias",
+	"2": "portComponent",
+	"3": "macAddress",
+	"4": "networkAddress",
+	"5": "interfaceName",
+	"6": "agentCircuitId",
+	"7": "local",
+}
+
+func normalizeLLDPSubtype(value string, mapping map[string]string) string {
+	if v, ok := mapping[value]; ok {
+		return v
+	}
+	return value
+}
+
 func decodeLLDPCapabilities(value string) []string {
 	bs, err := topologyutil.DecodeHexString(value)
 	if err != nil {
