--- src/go/plugin/go.d/collector/snmp_topology/topology_lldp_subtypes.go
@@ -1,23 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-package snmptopology
-
-var lldpChassisIDSubtypeMap = map[string]string{
-	"1": "chassisComponent",
-	"2": "interfaceAlias",
-	"3": "portComponent",
-	"4": "macAddress",
-	"5": "networkAddress",
-	"6": "interfaceName",
-	"7": "local",
-}
-
-var lldpPortIDSubtypeMap = map[string]string{
-	"1": "interfaceAlias",
-	"2": "portComponent",
-	"3": "macAddress",
-	"4": "networkAddress",
-	"5": "interfaceName",
-	"6": "agentCircuitId",
-	"7": "local",
-}
