+++ src/go/plugin/go.d/collector/snmp_topology/topology_management_address.go
@@ -3,11 +3,15 @@
 package snmptopology
 
 import (
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/collector/snmp_topology/internal/topologymodel"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/collector/snmp_topology/internal/topologyutil"
+	"encoding/hex"
+	"fmt"
 	"net"
 	"sort"
+	"strconv"
 	"strings"
+
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/collector/snmp_topology/internal/topologymodel"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/collector/snmp_topology/internal/topologyutil"
 )
 
 func normalizeAddressType(rawType, addr string) string {
@@ -124,4 +128,41 @@ func pickManagementIP(addrs []topologymodel.ManagementAddress) string {
 		return rawValues[0]
 	}
 	return ""
+}
+
+func reconstructLldpRemMgmtAddrHex(tags map[string]string) string {
+	lengthStr := strings.TrimSpace(tags[tagLldpRemMgmtAddrLen])
+	length, err := strconv.Atoi(lengthStr)
+	if err != nil || length <= 0 || length > net.IPv6len {
+		return ""
+	}
+
+	addr := make([]byte, 0, length)
+	for i := 1; i <= length; i++ {
+		tag := fmt.Sprintf("%s%d", tagLldpRemMgmtAddrOctetPref, i)
+		v := strings.TrimSpace(tags[tag])
+		if v == "" {
+			return ""
+		}
+		octet, err := strconv.Atoi(v)
+		if err != nil || octet < 0 || octet > 255 {
+			return ""
+		}
+		addr = append(addr, byte(octet))
+	}
+
+	return hex.EncodeToString(addr)
+}
+
+func normalizeManagementAddress(rawAddr, rawType string) (string, string) {
+	rawAddr = strings.TrimSpace(rawAddr)
+	if rawAddr == "" {
+		return "", normalizeAddressType(rawType, "")
+	}
+
+	if ip := topologyutil.NormalizeIPAddress(rawAddr); ip != "" {
+		return ip, normalizeAddressType(rawType, ip)
+	}
+
+	return rawAddr, normalizeAddressType(rawType, rawAddr)
 }
