+++ src/go/plugin/go.d/collector/snmp_topology/topology_metadata_test.go
